generic ui rendering
map/enum based

also,  preparatory work for a dialogue system:
- NPC class, an Entity specialization which contains a Conversation class
- Conversation class, which will store conversations as an n-ary tree
- Dialogue class for displaying the current Conversation
- useBuilding has been replaced by the "Interact" command, which finds the active interactable and then delegates to useBuilding or the new useNPC(), whichever is appropriate
- moved fade overlay to the renderer and cleaned up some of those semantics. game renderer now uses an enum-keyed map.

+++ include/game/game.hpp
@@ -5,6 +5,7 @@
 #include <map>
 #include <vector>
 
+#include <entity/npc.hpp>
 #include <entity/player.hpp>
 
 #include <item/library.hpp>
@@ -226,7 +227,7 @@ private:
     constexpr static float fade_threshold = 0.005f;
 
     sf::RectangleShape fade;
-    uint8_t fade_alpha;
+    int fade_alpha;
 
 /// draw ///
 ///
